AdventOfCode/AdventOfCode2024.Tests/Day03Tests.cs
namespace AdventOfCode2024.Tests;

public class Day03Tests : DayTest<Day03>
{
    private const string Input = "xmul(2,4)%&mul[3,7]!@^do_not_mul(5,5)+mul(32,64]then(mul(11,8)mul(8,5))";

    [Theory]
    [InlineData(Input, "161")]
    public override void Test1(string inputString, string expected)
    {
        var result = Day.Perform1(inputString);
        Assert.Equal(expected, result);
    }

    [Theory]
    [InlineData(Input, "4")]
    public override void Test2(string inputString, string expected)
    {
        var result = Day.Perform2(inputString);
        Assert.Equal(expected, result);
    }
}